pyproject.toml
[project]
name = "mytask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache[redis]>=0.12.3",
    "aioredis>=2.0.1",
    "async-substrate-interface>=1.1.0",
    "asyncpg>=0.30.0",
    "bittensor>=9.3.0",
    "celery>=5.5.1",
    "fastapi>=0.110.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
